+++ assets/src/ba_data/python/ba/_apputils.py
@@ -6,8 +6,12 @@ from __future__ import annotations
 import gc
 import os
 import logging
+from threading import Thread
+from dataclasses import dataclass
 from typing import TYPE_CHECKING
 
+from efro.log import LogLevel
+from efro.dataclassio import ioprepped, dataclass_to_json, dataclass_from_json
 import _ba
 
 if TYPE_CHECKING:
@@ -266,22 +270,67 @@ def print_corrupt_file_error() -> None:
 _tbfiles: list[TextIO] = []
 
 
-def dump_tracebacks(delay: float) -> None:
-    """Dump a traceback of all Python threads after a delay in seconds.
+@ioprepped
+@dataclass
+class DumpedAppStateMetadata:
+    """High level info about a dumped app state."""
 
-    Can be used for debugging deadlock situations. Will dump to a preset
-    file location in the app config dir. Will attempt to log and clear
-    the results after dumping. It will be done at next launch otherwise,
-    or can be done explicitly via log_dumped_tracebacks().
+    reason: str
+    app_time: float
+    log_level: LogLevel
 
-    Do not use this call during regular operation of the app; it is only
-    intended for debugging as it can leak file descriptors/etc.
+
+def dump_app_state(
+    delay: float = 0.0,
+    reason: str = 'Unspecified',
+    log_level: LogLevel = LogLevel.WARNING,
+) -> None:
+    """Dump various app state for debugging purposes.
+
+    This includes stack traces for all Python threads (and potentially
+    other info in the future).
+
+    This is intended for use debugging deadlock situations. It will dump
+    to preset file location(s) in the app config dir, and will attempt to
+    log and clear the results after dumping. If that should fail (due to
+    a hung app, etc.), then the results will be logged and cleared on the
+    next app run.
+
+    Do not use this call during regular smooth operation of the app; it
+    is should only be used for debugging or in response to confirmed
+    problems as it can leak file descriptors, cause hitches, etc.
     """
     # pylint: disable=consider-using-with
     import faulthandler
     from ba._generated.enums import TimeType
 
-    tbpath = os.path.join(os.path.dirname(_ba.app.config_file_path), '_tbdump')
+    # Dump our metadata immediately. If a delay is passed, it generally
+    # means we expect things to hang momentarily, so we should not delay
+    # writing our metadata or it will likely not happen. Though we
+    # should remember that metadata doesn't line up perfectly in time with
+    # the dump in that case.
+    try:
+        mdpath = os.path.join(
+            os.path.dirname(_ba.app.config_file_path), '_appstate_dump_md'
+        )
+        with open(mdpath, 'w', encoding='utf-8') as outfile:
+            outfile.write(
+                dataclass_to_json(
+                    DumpedAppStateMetadata(
+                        reason=reason,
+                        app_time=_ba.time(TimeType.REAL),
+                        log_level=log_level,
+                    )
+                )
+            )
+    except Exception:
+        # Abandon whole dump if we can't write metadata.
+        logging.exception('Error writing app state dump metadata.')
+        return
+
+    tbpath = os.path.join(
+        os.path.dirname(_ba.app.config_file_path), '_appstate_dump_tb'
+    )
 
     # faulthandler needs the raw file descriptor to still be valid when
     # it fires, so stuff this into a global var to make sure it doesn't get
@@ -289,23 +338,132 @@ def dump_tracebacks(delay: float) -> None:
     tbfile = open(tbpath, 'w', encoding='utf-8')
     _tbfiles.append(tbfile)
 
-    faulthandler.dump_traceback_later(delay, file=tbfile)
+    if delay > 0.0:
+        faulthandler.dump_traceback_later(delay, file=tbfile)
+    else:
+        faulthandler.dump_traceback(file=tbfile)
 
     # Attempt to log shortly after dumping.
-    with _ba.Context('ui'):
-        _ba.timer(delay + 1.0, log_dumped_tracebacks, timetype=TimeType.REAL)
+    # Allow sufficient time since we don't know how long the dump takes.
+    # We want this to work from any thread, so need to kick this part
+    # over to the logic thread so timer works.
+    _ba.pushcall(
+        lambda: _ba.timer(
+            delay + 1.0, log_dumped_app_state, timetype=TimeType.REAL
+        ),
+        from_other_thread=True,
+        suppress_other_thread_warning=True,
+    )
 
 
-def log_dumped_tracebacks() -> None:
-    """If a traceback dump exists, log it and clear it. No-op otherwise."""
+def log_dumped_app_state() -> None:
+    """If an app-state dump exists, log it and clear it. No-op otherwise."""
 
     try:
-        tbpath = os.path.join(
-            os.path.dirname(_ba.app.config_file_path), '_tbdump'
+        out = ''
+        mdpath = os.path.join(
+            os.path.dirname(_ba.app.config_file_path), '_appstate_dump_md'
         )
-        if os.path.exists(tbpath):
-            with open(tbpath, 'r', encoding='utf-8') as infile:
-                logging.info('Dumped tracebacks:\n%s', infile.read())
-            os.unlink(tbpath)
+        if os.path.exists(mdpath):
+            with open(mdpath, 'r', encoding='utf-8') as infile:
+                metadata = dataclass_from_json(
+                    DumpedAppStateMetadata, infile.read()
+                )
+            os.unlink(mdpath)
+            out += (
+                f'App state dump:\nReason: {metadata.reason}\n'
+                f'Time: {metadata.app_time:.2f}'
+            )
+            tbpath = os.path.join(
+                os.path.dirname(_ba.app.config_file_path), '_appstate_dump_tb'
+            )
+            if os.path.exists(tbpath):
+                with open(tbpath, 'r', encoding='utf-8') as infile:
+                    out += '\nPython tracebacks:\n' + infile.read()
+                os.unlink(tbpath)
+            logging.log(metadata.log_level.python_logging_level, out)
     except Exception:
-        logging.exception('Error logging dumped tracebacks.')
+        logging.exception('Error logging dumped app state.')
+
+
+class AppHealthMonitor:
+    """Logs things like app-not-responding issues."""
+
+    def __init__(self) -> None:
+        assert _ba.in_logic_thread()
+        self._running = True
+        self._thread = Thread(target=self._bg_thread_main, daemon=True)
+        self._thread.start()
+        self._response = False
+        self._first_check = True
+
+    def _bg_thread_main(self) -> None:
+
+        try:
+            self._monitor_app()
+        except Exception:
+            logging.exception('Error in AppHealthMonitor thread.')
+
+    def _set_response(self) -> None:
+        assert _ba.in_logic_thread()
+        self._response = True
+
+    def _check_running(self) -> bool:
+        # Workaround for the fact that mypy assumes _running
+        # doesn't change during the course of a function.
+        return self._running
+
+    def _monitor_app(self) -> None:
+        import time
+
+        while bool(True):
+
+            # Always sleep a bit between checks.
+            time.sleep(1.234)
+
+            # Do nothing while backgrounded.
+            while not self._running:
+                time.sleep(2.3456)
+
+            # Wait for the logic thread to run something we send it.
+            starttime = time.monotonic()
+            self._response = False
+            _ba.pushcall(self._set_response, raw=True)
+            while not self._response:
+
+                # Abort this check if we went into the background.
+                if not self._check_running():
+                    break
+
+                # Wait a bit longer the first time through since the app
+                # could still be starting up; we generally don't want to
+                # report that.
+                threshold = 10 if self._first_check else 5
+
+                # If we've been waiting too long (and the app is running)
+                # dump the app state and bail. Make an exception for the
+                # first check though since the app could just be taking
+                # a while to get going; we don't want to report that.
+                duration = time.monotonic() - starttime
+                if duration > threshold:
+                    dump_app_state(
+                        reason=f'Logic thread unresponsive'
+                        f' for {threshold} seconds.'
+                    )
+
+                    # We just do one alert for now.
+                    return
+
+                time.sleep(1.042)
+
+            self._first_check = False
+
+    def on_app_pause(self) -> None:
+        """Should be called when the app pauses."""
+        assert _ba.in_logic_thread()
+        self._running = False
+
+    def on_app_resume(self) -> None:
+        """Should be called when the app resumes."""
+        assert _ba.in_logic_thread()
+        self._running = True