+++ assets/src/ba_data/python/ba/internal.py
@@ -101,8 +101,8 @@ from ba._apputils import (
     is_browser_likely_available,
     get_remote_app_name,
     should_submit_debug_info,
-    dump_tracebacks,
-    log_dumped_tracebacks,
+    dump_app_state,
+    log_dumped_app_state,
 )
 from ba._benchmark import (
     run_gpu_benchmark,
@@ -334,6 +334,6 @@ __all__ = [
     'sign_out_v1',
     'sign_in_v1',
     'mark_config_dirty',
-    'dump_tracebacks',
-    'log_dumped_tracebacks',
+    'dump_app_state',
+    'log_dumped_app_state',
 ]