+++ src/ballistica/ballistica.cc
@@ -32,8 +32,8 @@
 namespace ballistica {
 
 // These are set automatically via script; don't modify them here.
-const int kAppBuildNumber = 20967;
-const char* kAppVersion = "1.7.16";
+const int kAppBuildNumber = 20973;
+const char* kAppVersion = "1.7.17";
 
 // Our standalone globals.
 // These are separated out for easy access.